third_party/blink/web_tests/platform/generic/fast/dom/StyleSheet/detached-style-pi-expected.txt

Bug 45205: Test that <xml-stylesheet> processing instructions only have stylesheets when in document.

PASS successfullyParsed is true

TEST COMPLETE

PASS sheet.ownerNode is pi
PASS pi.sheet === sheet is true
Removing <pi>...
PASS pi.sheet is null
PASS sheet.ownerNode is null
PASS sheet.cssRules[0].cssText[0] is "*"
Re-adding <pi>...
PASS sheet.ownerNode is null
PASS pi.sheet === sheet is false
PASS pi.sheet.ownerNode is pi
